requirements.txt
flask==3.1.0
openai==1.76.0
python-dotenv==1.1.0
werkzeug==3.1.3